+++ src/lib/st-form/st-form.component.ts
@@ -72,6 +72,9 @@ export class StFormComponent implements AfterViewInit, AfterViewChecked, Control
     */
    @Input() sectionDescriptionLevel: number = -1;
 
+   /** @Input {number} [textFieldMaxWidth=] Maximum width of a field needed to paint a input or textarea */
+   @Input() textFieldMaxWidth: number;
+
    /** @Input {boolean} [showTooltips=-1] Enable or disable displaying of tooltips
     * By default, tooltips are displayed
     */