+++ src/lib/st-form/st-form.module.ts
@@ -19,6 +19,7 @@ import { StFormDirectiveModule } from '../directives/form/form-directives.module
 import { StFormFieldModule } from './st-form-field/st-form-field.module';
 import { StSelectModule } from '../st-select/st-select.module';
 import { StSwitchModule } from '../st-switch/st-switch.module';
+import { StTextareaModule } from '../st-textarea/st-textarea.module';
 
 @NgModule({
    imports: [
@@ -30,7 +31,8 @@ import { StSwitchModule } from '../st-switch/st-switch.module';
       PipesModule,
       StSelectModule,
       StSwitchModule,
-      StFormDirectiveModule
+      StFormDirectiveModule,
+      StTextareaModule
    ],
    declarations: [StFormComponent],
    exports: [StFormComponent]