+++ src/lib/st-textarea/st-textarea.error.model.ts
@@ -11,11 +11,14 @@
 import { TranslateableElement } from '../utils/egeo-resolver/egeo-resolve-model';
 
 export interface StTextareaError {
-  generic?: string;
-  required?: string;
+   generic?: string;
+   required?: string;
+   minLength?: string;
+   maxLength?: string;
+   pattern?: string;
 }
 
 export interface StTextareaErrorSchema {
-  generic?: TranslateableElement;
-  required?: TranslateableElement;
+   generic?: TranslateableElement;
+   required?: TranslateableElement;
 }